+++ TDEngine2/include/graphics/CBaseMaterial.h
@@ -371,6 +371,8 @@ namespace TDEngine2
 			*/
 
 			TDE2_API ITexture* GetTextureResource(const std::string& id, TMaterialInstanceId instanceId = DefaultMaterialInstanceId) const override;
+
+			TDE2_API bool IsScissorTestEnabled() const override;
 		protected:
 			DECLARE_INTERFACE_IMPL_PROTECTED_MEMBERS(CBaseMaterial)