+++ TDEngine2/include/graphics/IMaterial.h
@@ -345,6 +345,8 @@ namespace TDEngine2
 			*/
 
 			TDE2_API virtual ITexture* GetTextureResource(const std::string& id, TMaterialInstanceId instanceId = DefaultMaterialInstanceId) const = 0;
+
+			TDE2_API virtual bool IsScissorTestEnabled() const = 0;
 		protected:
 			DECLARE_INTERFACE_PROTECTED_MEMBERS(IMaterial)