+++ TDEngine2/plugins/D3D11GraphicsContext/include/CD3D11GraphicsContext.h
@@ -115,7 +115,7 @@ namespace TDEngine2
 				\param[in] scissorRect A rectangle for scissor test
 			*/
 
-			TDE2_API void SetScissorRect(const TRectF32& scissorRect) override;
+			TDE2_API void SetScissorRect(const TRectU32& scissorRect) override;
 
 			/*!
 				\brief The method computes a perspective projection matrix specific for a graphics context